[feature/UserViewModel]refactor: Refactor move joinMember func LoginViewModel -> UserViewModel
